migrate previous blueprints
Migrate the blueprints in the storj/storj repository to this repository.

The migration involved renaming the files using the blueprint title and
the Git authored date, adding a front matter, and arranging the names of
the assets to match the doc name (updating their references inside of
the document).

It doesn't add any tag to the front matter and archives any blueprint
that doesn't apply anymore.

This commit also readdress the README, updating the conventions and
adding the exceptions that these imported documents have.